Refactor LandingPage and Navbar components
- Removed the ResizableNavbar component and its associated logic from LandingPage for a cleaner structure.
- Introduced NavbarWrapper in the RootLayout to encapsulate navigation functionality.
- Updated NavItemsProps to allow React nodes for item names, enhancing flexibility in rendering navigation items.
